+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/metadata/expressions/FieldKeyExpression.java
@@ -188,6 +188,7 @@ public class FieldKeyExpression extends BaseKeyExpression implements AtomKeyExpr
         return RecordMetaDataProto.Field.newBuilder()
                 .setFieldName(fieldName)
                 .setFanType(fanType.toProto())
+                .setNullInterpretation(nullStandin.toProto())
                 .build();
     }
 
@@ -295,6 +296,10 @@ public class FieldKeyExpression extends BaseKeyExpression implements AtomKeyExpr
     @Nonnull
     public FanType getFanType() {
         return fanType;
+    }
+
+    public Key.Evaluated.NullStandin getNullStandin() {
+        return nullStandin;
     }
 
     @Override